\name{botDividendsUS}
\alias{botDividendsUS}
\title{botDividendsUS trae los dividendos de activos para la función homonima de testBot.}
\usage{
botDividendsUS()
}
\description{
Deja el archivo "dividendsUS.xlsx" en ~/Downloads/temp/ para que el bot lo levante y lo devuelva.
}
\examples{
botDividendsUS()
}
