% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbot-package.R
\docType{package}
\name{testbot-package}
\alias{testbot-package}
\title{testbot: Bring package test coverage to 100}
\description{
This package uses an LLM to inspect existing package test code and write new code to bring test coverage to 100%. This is particularly useful for packages that contain compiled code, to ensure that all code paths are hit and diagnostic tools such as Valgrind, ASAN/UBSAN give more reliable results.
}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
