% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbot.R
\name{testbot}
\alias{testbot}
\title{Testbot}
\usage{
testbot(client = NULL, retries = 2L)
}
\arguments{
\item{client}{the default \code{NULL} uses \code{\link[ellmer:chat_anthropic]{ellmer::chat_anthropic()}}, but you may
pass an alternative such as \code{\link[ellmer:chat_openai]{ellmer::chat_openai()}}.}

\item{retries}{integer number of times to retry if an error occurs mid-way.
This may happen due to timeout or for other stochastic reasons.}
}
\value{
An ellmer 'Chat' object.
}
\description{
For packages, creates an additional test file "tests/testthat/test-bot.R" to
bring test coverage to 100\\%.
}
\details{
Call this function from the top working directory of your package.
}
